walker_python/container_full_backup/requirements.txt
# Core dependencies
numpy>=1.21.0
box2d-py>=2.3.8
pygame>=2.0.0

# Web interface
flask>=2.0.0
flask-socketio>=5.0.0
websockets>=10.0

# Data science and visualization
matplotlib>=3.5.0
pandas>=1.3.0
scipy>=1.7.0

# ML experiment tracking
mlflow>=3.1.0

# System monitoring
psutil>=5.8.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0 